refactor: add redis-logger.ts
