+++ package.json
@@ -113,10 +113,7 @@
             "@/(.*)": [
                 "<rootDir>/$1"
             ]
-        },
-        "coverageReporters": [
-            "json-summary"
-        ]
+        }
     },
     "commitlint": {
         "extends": [
